+++ EcommerceAppAPI/src/main/java/com/project/EcommerceAppAPI/repositories/ProductCategoryRepository.java
@@ -3,6 +3,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.project.EcommerceAppAPI.models.ProductCategory;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface ProductCategoryRepository extends JpaRepository<ProductCategory,Integer> {
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory,Long> {
 }
